stubgen script
